Better way to export actions
