+++ src/__tests__/index.js
@@ -5,24 +5,15 @@
  no-unused-expressions
 */
 import { expect } from 'chai';
-import { fromJS } from 'immutable';
-import { startPlugin } from '..';
-import { mockSocket } from '../socket/__mocks__';
+import { startPlugin, actions } from '..';
 
-describe('startPlugin', function() {
-  it('returns the store', function() {
-    const plugin = startPlugin({});
-    expect(plugin.getStore()).to.not.be.null;
+describe('index', function() {
+  it('exports startPlugin', function() {
+    expect(startPlugin).to.be.a('function');
   });
 
-  it('authenticates if auth exists', function() {
-    const workerUUID = 'abc123';
-    const socketAuth = { auth: 'SEKRET', sig: 'SEKRETSIG' };
-    const auth = { workerUUID, socketAuth };
-
-    const plugin = startPlugin({ auth, socketConstructor: mockSocket() });
-    const state = plugin.getStore().getState();
-    expect(state.worker.get('uuid')).to.equal(workerUUID);
-    expect(state.socket.get('auth')).to.equal(fromJS(socketAuth));
+  it('exports actions', function() {
+    expect(actions.assignWork).to.be.a('function');
+    expect(actions.authenticate).to.be.a('function');
   });
 });
